Table/Affinities
Skill/Enemy/Trait table number offsets have been updated to how they would look like in other parts of the file.
Universal Power has been added.
Static Electricity has been removed as it is not in NAME.TBL.
Several previously unknown affinities have been roughly filled in by putting the number for it on an element for Arsene. The recorded affinity is just from the Persona status screen, not sure if it is fully accurate or if there is more there.